refactor: 데이터, 유틸리티 함수, 알림 객체 배열 파일 분리 -types.ts: 타입 정의 data/alarms.ts: Alarms 배열 데이터 utils/timeUtils.ts: 시간 관련 유틸리티 함수 components/AlarmItem.tsx: AlarmItem 컴포넌트 pages/AlarmPage.tsx: 메인 AlarmPage 컴포넌트
