+++ src/data/types.ts
@@ -1,0 +1,8 @@
+export interface Alarm {
+    id: number;
+    title: string;
+    about: string;
+    tag: string;
+    time: number;
+    isRead?: boolean;
+  }
